refactor: PoCで利用するtypesの保存場所を変更
